pyproject.toml
[project]
name = "fastapi_app"
version = "0.1.0"
description = ""
authors = [
    { name = "Remist9", email = "serega189ser@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn>=0.34.1,<0.35.0",
    "sqlalchemy>=2.0.40,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "python-jose[cryptography]>=3.4.0,<4.0.0",
    "python-dotenv>=1.1.0,<2.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

